fix: test shared class DebugUtils and DisposableList move Shared directory.
