--- test/PipingServer.Core.Tests/DebugUtils.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Diagnostics;
-using System.Threading;
-internal static class DebugUtils
-{
-    public static CancellationTokenSource CreateTokenSource(TimeSpan delay)
-    {
-        return Debugger.IsAttached
-            ? new CancellationTokenSource()
-            : new CancellationTokenSource(delay);
-    }
-}
